Rename generator to writer
