+++ gcovr/tests/test_html_generator.py
@@ -7,7 +7,7 @@
 #
 # _____________________________________________________________________________
 #
-# Copyright (c) 2013-2020 the gcovr authors
+# Copyright (c) 2013-2021 the gcovr authors
 # Copyright (c) 2013 Sandia Corporation.
 # This software is distributed under the BSD License.
 # Under the terms of Contract DE-AC04-94AL85000 with Sandia Corporation,
@@ -19,7 +19,7 @@
 import os
 import pytest
 import sys
-from ..generator.html import _make_short_sourcename
+from ..writer.html import _make_short_sourcename
 
 CurrentDrive = os.getcwd()[0:1]
